fix(arkruntime): support fractional task duration

+++ service/arkruntime/model/content_generation.go
@@ -92,7 +92,7 @@ type GetContentGenerationTaskResponse struct {
 	Resolution            *string                  `json:"resolution,omitempty"`
 	Ratio                 *string                  `json:"ratio,omitempty"`
 	OutputFormat          *string                  `json:"output_format,omitempty"`
-	Duration              *int64                   `json:"duration,omitempty"`
+	Duration              *float64                 `json:"duration,omitempty"`
 	CreatedAt             int64                    `json:"created_at"`
 	UpdatedAt             int64                    `json:"updated_at"`
 	Seed                  *int64                   `json:"seed,omitempty"`
